fix increment predictions output, add script to clean summary evaluations
in the increment table function, when new predictions were being appended it was capturing the indices of the old and new csvs, causing extra columns to be recorded. I've specified the columns we want to return to fix this issue

in the summary_results, there were duplicates for Friday if the script captured dates on the weekend.  drop_dups.py removes the duplicates